+++ demo/meeting/src/containers/DevicePermissionControl/Styled.tsx
@@ -1,8 +1,0 @@
-// Copyright 2021 Amazon.com, Inc. or its affiliates. All Rights Reserved.
-// SPDX-License-Identifier: Apache-2.0
-
-import styled from 'styled-components';
-
-export const StyledP = styled.p`
-  padding: 1rem 1rem 1rem 0;
-`;